+++ package.json
@@ -88,7 +88,6 @@
   },
   "sideeffects": false,
   "dependencies": {
-    "@types/mathjs": "^5.0.1",
     "mathjs": "^5.9.0",
     "vue-test-utils": "^1.0.0-beta.11"
   }
