+++ tsconfig.json
@@ -13,7 +13,8 @@
     "baseUrl": ".",
     "types": ["webpack-env", "jest"],
     "paths": {
-      "@/*": ["src/*"]
+      "@/*": ["src/*"],
+      "*": ["src/typings/*"]
     },
     "lib": ["es5", "esnext", "dom", "dom.iterable", "scripthost"]
   },
